==> 03-preprocessing.ipynb <==
### CONSTANTS

DATA_PATH = "data/data_feature_engineered.csv"
NUMERICAL_FEATURES = ['Air temperature [K]', 'Process temperature [K]', 'Rotational speed [rpm]',
                       'Torque [Nm]', 'Tool wear [min]', 'feature_1', 'feature_2', 'feature_3', 'feature_4']
CATEGORICAL_FEATURES = ['Type_M', 'Type_L', 'Type_H']
TARGET = "Target"

### Importing Libraries

import numpy as np
import pandas as pd

from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from category_encoders import OneHotEncoder

import pickle

### Importing Data

data = pd.read_csv(DATA_PATH)
data.head()

### 01. Data Encoding

ohe = OneHotEncoder(use_cat_names=True)
encoded_columns = ohe.fit_transform(data['Type'])
data = pd.concat([data, encoded_columns], axis=1).drop(columns="Type")
data.head()

### 02. Data Splitting

X = data.drop(columns=TARGET)
y = data[TARGET]
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, stratify=y)
print(f'X_train shape: {X_train.shape}')
print(f'X_test shape: {X_test.shape}')
print(f'y_train shape: {y_train.shape}')
print(f'y_test shape: {y_test.shape}')

### 03. Fit Data Scaler

ss = StandardScaler()
ss.fit(X_train[NUMERICAL_FEATURES])

### 04. Exporting Preprocessed Data and Scaler

X_train.to_csv("data/X_train.csv", index=False)
X_test.to_csv("data/X_test.csv", index=False)
y_train.to_csv("data/y_train.csv", index=False)
y_test.to_csv("data/y_test.csv", index=False)

pickle.dump(ss, open("models/standard_scaler.pickle.dat", "wb"))
