==> 06-DNN.ipynb <==
### CONSTANTS

PATHES = ['data/X_train.csv', 'data/X_test.csv', 'data/y_train.csv', 'data/y_test.csv']

### Importing Libraries

import numpy as np
import pandas as pd

import torch
from torch import nn
from torch.utils.data import TensorDataset, DataLoader
from sklearn.metrics import classification_report
from imblearn.over_sampling import RandomOverSampler

from tqdm import tqdm

### Functions

def get_train_test_data(pathes):
    data = []
    for path in pathes:
        df = pd.read_csv(path)
        data.append(df)
    return data

def train_step(model, data_loader, loss_fn, optimizer):
    train_loss, train_acc = 0, 0
    model.train()
    for X, y in data_loader:
        y_logit = model(X)

        loss = loss_fn(y_logit, y)
        train_loss += loss
        
        optimizer.zero_grad()

        loss.backward()

        optimizer.step()

        predicted_classes = torch.sigmoid(y_logit).round()
        train_acc += (predicted_classes == y).sum().item()/len(predicted_classes)
    
    train_loss /= len(data_loader)
    train_acc /= len(data_loader)
    return train_loss, train_acc

def test_step(model, data_loader, loss_fn):
    test_loss, test_acc = 0, 0
    model.eval()
    with torch.inference_mode():
        for X, y in data_loader:
            y_logit = model(X)

            test_loss += loss_fn(y_logit, y)

            predicted_classes = torch.sigmoid(y_logit).round()
            test_acc += (predicted_classes == y).sum().item()/len(predicted_classes)

        test_loss /= len(data_loader)
        test_acc /= len(data_loader)
        return test_loss, test_acc
    
def train(model, train_dataloader, test_dataloader, loss_fn, optimizer, epochs):
    result = {
        'train_loss': [],
        'train_acc': [],
        'test_loss': [],
        'test_acc': []
    }
    print("=================================================")

    for epoch in tqdm(range(epochs)):
        train_loss, train_acc = train_step(model,
                                           train_dataloader,
                                           loss_fn,
                                           optimizer)
        
        test_loss, test_acc = test_step(model,
                                        test_dataloader,
                                        loss_fn)
        
        result['train_loss'].append(train_loss.item())
        result['train_acc'].append(train_acc)
        result['test_loss'].append(test_loss.item())
        result['test_acc'].append(test_acc)
        if epoch%10==1:
            print(f" Epoch: {epoch+1}")
            print(f"Train Loss = {train_loss:.4f} || Train Accuracy = {train_acc*100:.2f}%")
            print(f"Test Loss = {test_loss:.4f} || Test Accuracy = {test_acc*100:.2f}%")
            print("=================================================")
        
    return result

### Reading Data

X_train, X_test, y_train, y_test = get_train_test_data(PATHES)
over_sampler = RandomOverSampler()
X_train, y_train = over_sampler.fit_resample(X_train, y_train)
print(f'X_train shape: {X_train.shape}')
print(f'X_test shape: {X_test.shape}')
print(f'y_train shape: {y_train.shape}')
print(f'y_test shape: {y_test.shape}')

### Creating Datasets & Dataloaders

X_train = torch.tensor(X_train.values, dtype=torch.float32)
X_test = torch.tensor(X_test.values, dtype=torch.float32)
y_train = torch.tensor(y_train.values, dtype=torch.float32)
y_test = torch.tensor(y_test.values, dtype=torch.float32)

train_data = TensorDataset(X_train, y_train)
test_data = TensorDataset(X_test, y_test)

train_loader = DataLoader(train_data, batch_size=256, shuffle=True, drop_last=True)
test_loader = DataLoader(test_data, batch_size=256, shuffle=True, drop_last=True)

### Model V1

class ModelV1(nn.Module):
    def __init__(self, input_shape, hidden_units, output_shape):
        super().__init__()
        self.input_layer = nn.Sequential(
            nn.Linear(input_shape, hidden_units),
            nn.ReLU()
        )
        self.hidden_block = nn.Sequential(
            nn.Linear(hidden_units, 4*hidden_units),
            nn.ReLU(),
            nn.Linear(4*hidden_units, 2*hidden_units),
            nn.ReLU(),
            nn.Linear(2*hidden_units, hidden_units),
            nn.ReLU(),
            nn.Linear(hidden_units, int(0.5*hidden_units)),
            nn.ReLU()
        )
        self.classifier = nn.Linear(int(0.5*hidden_units), output_shape)
    
    def forward(self, x):
        x = self.input_layer(x)
        x = self.hidden_block(x)
        x = self.classifier(x)
        return x


### Initiating Model, defining loss function and optimizer

model = ModelV1(12, 16, 1)
model

loss_fn = nn.BCEWithLogitsLoss()
optimizer = torch.optim.Adam(model.parameters(), lr=0.001)

### Building Training and Testing loops

train(model, train_loader, test_loader, loss_fn, optimizer, epochs=50)

with torch.inference_mode():
        true_ys = []
        pred_ys = []
        for X, y in test_loader:
            true_ys.append(y)
            y_logit = model(X)
            predicted_classes = torch.sigmoid(y_logit).round()
            pred_ys.append(predicted_classes)

y_pred = torch.concat(pred_ys).squeeze().numpy()
y_true = torch.concat(true_ys).squeeze().numpy()

print(classification_report(y_true, y_pred))
